RPC: Fully remove delay between ago event reception and responding of pending getEvent calls
On incoming message, we wake up the mg_servers with the mg_wakeup_server
call. This will poll all pending getEvent connections, and it will notice the new event.
Most noticeable is that the ago UI buttons now respond directly when clicked on,
compared to previous up to 1s delay.