Replace pthread mutex with boost mutex, and RAII style locks to prevent missed unlocks.
RAII style locking means that the guard_lock object locks the mutex in the
constructor, and unlocks it in the destructor.
This means we do not have to think about manually unlocking it, neither
in normal code flow or when exceptions happens.

http://www.boost.org/doc/libs/1_56_0/doc/html/thread/synchronization.html#thread.synchronization.tutorial.internal_locking.internal_locking

This should work with Boost 1.49 as well; testing in agorpc only first.