resolver/UI: add a proper "delete device" command (and fix a segfault)
prev code emitted a "device.remove" event, and waited for response
(Which was never sent), and then a setdevicename with blank. Took some
time and gave confusing UI errors.
Added explicit deletedevice command instead, which does the same.

Also replaced some mutating checks for existing uuid in inventory, which
created blank inventory entries (device.remove trigged this, among other
things).
If a device was re-added through discovery after this, it would be
re-added as a device, but without a "stale" entry.. which later trigged
a crash in staleFunction when it tried to do asUint() on this
non-existing entry.